+++ src/Hooks_oCAIHuman.hpp
@@ -30,8 +30,8 @@ namespace GOTHIC_NAMESPACE
             }
         }
     }
-    auto PartialHook__oCAIHuman_CheckActiveSpells = ::Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x0046c2d0, 0x00474680, 0x00470fd8, 0x00472128)), &oCAIHuman_CheckActiveSpells);
-
+    auto PartialHook__oCAIHuman_CheckActiveSpells = ::Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x0046c2c8, 0x00474678, 0x00470fd8, 0x00472128)), &oCAIHuman_CheckActiveSpells);
+    
     void __fastcall oCAIHuman_MagicCheckSpellStates_RemoveScroll(::Union::Registers& reg)
     {
         auto magBook = reinterpret_cast<oCMag_Book*>(reg.esi);
@@ -45,14 +45,14 @@ namespace GOTHIC_NAMESPACE
             return;
         }
 
-        if(spellData->GetType() != oCSpell_Data::oCSpell_Type::SPELL_TYPE_TELEKINESIS)
+        if(!spellData->GetIsInvestSpell())
         {
             return;
         }
 
         auto spellItem = magBook->spellitems[spellNr];
 
-        if(!spellItem->MultiSlot() || CurrentTelStep < 4)
+        if(!spellItem->MultiSlot())
         {
             return;
         }
@@ -62,7 +62,7 @@ namespace GOTHIC_NAMESPACE
             return;
         }
 
-        auto item = ownerNpc->RemoveFromInv(magBook->spellitems[spellNr], 0);
+        auto item = ownerNpc->RemoveFromInv(spellItem, 0);
         magBook->wld->RemoveVob(item);
         
         if(magBook->spells.GetNum() > 0)
@@ -82,4 +82,26 @@ namespace GOTHIC_NAMESPACE
         }
     }
     auto PartialHook__oCAIHuman_MagicCheckSpellStates_RemoveScroll = ::Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x0046ced6, 0x0047533c, 0x00471bcb, 0x00472d0d)), &oCAIHuman_MagicCheckSpellStates_RemoveScroll);
+    
+    /*void __fastcall oCAIHuman_MagicMode_RemoveScroll(::Union::Registers& reg)
+    {
+        auto spell = reinterpret_cast<oCSpell*>(reg.ecx);
+
+        auto spellData = sdManager->GetSpellData(spell->spellID);
+        if(!spellData)
+        {
+            return;
+        }
+
+        if(!spellData->GetIsInvestSpell())
+        {
+            return;
+        }
+
+        reg.eax = 0x1;
+    }
+    auto PartialHook__oCAIHuman_MagicMode_RemoveScroll = ::Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x0046d3be, 0x0047586d, 0x00472037, 0x00473159)), &oCAIHuman_MagicMode_RemoveScroll);
+#if ENGINE > Engine_G2
+    auto PartialHook__oCAIHuman_MagicMode_RemoveScroll_G2 = ::Union::CreatePartialHook(reinterpret_cast<void*>(zSwitch(0x0, 0x0, 0x004721e6, 0x00473308)), &oCAIHuman_MagicMode_RemoveScroll);
+#endif*/
 }