+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+[1.0.2]
+- Fixed some addresses
+- Added new field to `C_SPELL_DATA` class: `spellIsInvestSpell`
+  - For spells like telekinesis, spell scroll will be removed from inventory after investing mana
+
 [1.0.1]
 - A bit of cleaning in the code
 - Fix for removing telekinesis scroll